refactor: main.py를 create_app() 팩토리로 축소, startup 로직을 core/startup으로 분리
- run_migrations/시그널 핸들러/플러그인 초기화/Docker 이미지 pull을
  app/core/startup.py로 이동 (로직 변경 없음)
- run_migrations는 import 부수효과에서 create_app() 내 명시 호출로 전환
- main.py 337줄 → 51줄, uvicorn 진입점(app.main:app) 불변
